added pages for restaurants viewing added a seller dashboard for seller to add new restaurant and new pizza reconciled package.json modified authenticated layout to provide a navbar minor adjustments to pizzas

+++ app/Http/Controllers/PizzaController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\Pizza;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use App\Models\User;
 
 class PizzaController extends Controller
 {
@@ -35,7 +37,26 @@ class PizzaController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        if (!$request->user) {
+            abort(401, "You must be logged in!");
+        }
+        $user = User::find($request->user);
+        if (!$user || !$user->is_seller) {
+            abort(401, "You must be a seller to edit or delete data");
+        }
+        $request->validate([
+            'name' => 'required',
+            'image' => 'required',
+            'ingredients' => 'required'
+        ]);
+
+        $pizza = Pizza::create([
+            'name' => $request->name,
+            'image' => $request->image,
+            'ingredients' => $request->ingredients
+        ]);
+
+        return response()->json($pizza);
     }
 
     /**
@@ -75,19 +96,21 @@ class PizzaController extends Controller
      */
     public function update(Request $request, Pizza $pizza, $key)
     {
-        try {
-            $piz = Pizza::find(intval($key));
-            $piz->update([
-                'name' => $request->name,
-                'image' => $request->image,
-                'ingredients' => $request->ingredients
-            ]);
-
-            return response()->json($piz);
-        } catch (\Throwable $th) {
-            abort(404, "Resource not found");
+        $user_id = $request->user;
+        if (!$user_id) {
+            abort(401, "You must be logged in!");
         }
-        
+        $user = User::find($user_id);
+        if (!$user->is_seller) {
+            abort(401, "You must be a seller to edit or delete data");
+        }
+        $piz = Pizza::find(intval($key));
+        $piz->update([
+            'name' => $request->name,
+            'image' => $request->image,
+            'ingredients' => $request->ingredients
+        ]);
+        return response()->json($piz);
     }
 
     /**
@@ -96,16 +119,18 @@ class PizzaController extends Controller
      * @param  \App\Models\Pizza  $pizza
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Pizza $pizza, $key)
+    public function destroy(Pizza $pizza, $key, Request $request)
     {
-        try {
-            $piz = Pizza::find(intval($key));
-            $piz->delete();
-            return response()->json($piz);
-            
-        } catch (\Throwable $th) {
-            abort(404, "Resource not found");
+        $user_id = $request->user;
+        if (!$user_id) {
+            abort(401, "You must be logged in!");
         }
-        
+        $user = User::find($user_id);
+        if (!$user->is_seller) {
+            abort(401, "You must be a seller to edit or delete data");
+        }
+        $piz = Pizza::find(intval($key));
+        $piz->delete();
+        return response()->json($piz);
     }
 }
